pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "replayci"
version = "0.1.0"
description = "A V1 foundation for ReplayCI developer tool"
authors = [{name = "Trae AI", email = "trae@example.com"}]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
]
requires-python = ">=3.11"

[project.scripts]
replayci = "replayci.cli:app"

[tool.setuptools.packages.find]
include = ["replayci*"]
